plan: split slot assignment and conv-param resolution out of the builder
Two phases lift cleanly out of the 1779-line plan builder, each depending
only on the graph and neither on anything the fusion scan decides.

assign_slots interns every tensor name to a dense id — which is what lets
the per-inference environment index a Vec instead of hashing strings —
pre-resolves each node's inputs and outputs to those ids, classifies
operator types once into NodeKind, and records use counts so the runner
can free a tensor as soon as its last reader has run. It also assigns the
tower-parallel branch tags; that is scheduling rather than indexing, but
it depends only on the graph, so splitting it into a third sweep would
buy nothing.

resolve_conv_params reads strides, pads, group and the
depthwise/pointwise classification once, keeping attribute lookups out of
the per-inference Conv dispatch. Its doc comment records why it is
awkward: it has to interpret the weight's shape against whichever of four
layouts the loader pre-permuted it into — the same leak ir::WeightLayout
works around, and the one a later phase is meant to remove by moving the
permute into prepack_weights.

build.rs drops from 1779 to 1550 lines. What remains is the fusion scan,
its post-passes and weight prepacking — deliberately left whole, because
select_fusions is about to be rewritten onto def-use matching and
splitting it first would mean splitting code that is then replaced.

No logic changed; the phases are the same statements in the same order,
with their outputs named. The bodies moved verbatim apart from
resolve_conv_params gaining node_kinds as a parameter, which it
previously read from the enclosing scope.

Validation: 233 yscv-onnx tests, 2305 workspace, unchanged. cargo fmt,
clippy --workspace --all-targets --all-features -D warnings,
check-doc-counts.sh.
